refactor: src/validations.ts -> src/diagnostics.ts
