+++ src/index.ts
@@ -6,4 +6,4 @@ export * from './logs';
 export * from './parsers';
 export * from './schemas';
 export * from './servers';
-export * from './validations';
+export * from './diagnostics';
